Rename tokens.rs to token_admin_registry.rs
