+++ chains/solana/contracts/programs/ccip-router/src/instructions/v1/mod.rs
@@ -4,7 +4,7 @@
 pub mod admin;
 pub mod offramp;
 pub mod onramp;
-pub mod tokens;
+pub mod token_admin_registry;
 
 /////////////////////
 // Private modules //
