+++ chains/solana/contracts/programs/ccip-router/src/instructions/v1/token_admin_registry.rs
